+++ deps/v8/src/snapshot/snapshot-source-sink.h
@@ -50,7 +50,7 @@ class SnapshotByteSource final {
 
   void Advance(int by) { position_ += by; }
 
-  void CopyRaw(void* to, int number_of_bytes) {
+  void CopyRaw(void* to, size_t number_of_bytes) {
     DCHECK_LE(position_ + number_of_bytes, length_);
     memcpy(to, data_ + position_, number_of_bytes);
     position_ += number_of_bytes;
@@ -108,16 +108,16 @@ class SnapshotByteSource final {
   // Returns length.
   int GetBlob(const uint8_t** data);
 
-  int position() const { return position_; }
-  void set_position(int position) { position_ = position; }
+  size_t position() const { return position_; }
+  void set_position(size_t position) { position_ = position; }
 
   const uint8_t* data() const { return data_; }
-  int length() const { return length_; }
+  size_t length() const { return length_; }
 
  private:
   const uint8_t* data_;
-  int length_;
-  int position_;
+  size_t length_;
+  size_t position_;
 };
 
 /**
@@ -134,17 +134,17 @@ class SnapshotByteSink {
 
   void Put(uint8_t b, const char* description) { data_.push_back(b); }
 
-  void PutN(int number_of_bytes, const uint8_t v, const char* description);
+  void PutN(size_t number_of_bytes, const uint8_t v, const char* description);
   // Append a uint30 with run-length encoding. Must be decoded with GetUint30.
   void PutUint30(uint32_t integer, const char* description);
   void PutUint32(uint32_t integer, const char* description) {
     PutRaw(reinterpret_cast<uint8_t*>(&integer), sizeof(integer), description);
   }
-  void PutRaw(const uint8_t* data, int number_of_bytes,
+  void PutRaw(const uint8_t* data, size_t number_of_bytes,
               const char* description);
 
   void Append(const SnapshotByteSink& other);
-  int Position() const { return static_cast<int>(data_.size()); }
+  size_t Position() const { return data_.size(); }
 
   const std::vector<uint8_t>* data() const { return &data_; }
